requirements.txt
streamlit==1.43.0
requests==2.32.3
openai==1.65.2
pillow==11.1.0
